models/yolo/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
ultralytics==8.0.199
pillow==10.0.0
numpy==1.26.0
opencv-python-headless==4.8.1.78

# If you run on GPU and want optimized torch, uncomment the matching packages below
# (ensure the base image has the right CUDA/runtime):
# torch==2.2.0+cu118
# torchvision==0.15.1+cu118
# torchaudio==2.2.0+cu118

# Note: Some opencv builds are published with different build tags; if you see
# "No matching distribution found" during pip install, try a slightly different
# patch-level (e.g. 4.8.1.78) or remove the patch suffix to allow pip to pick
# a compatible wheel for the target Python version.